TECH: Refactor connector hooks
